requirements.txt
fastapi
uvicorn
python-dotenv
httpx
beautifulsoup4
aiohttp
wikipedia
sqlalchemy
psycopg2-binary
python-multipart
jinja2
pydantic-settings
python-jose
passlib
bcrypt
langchain
langchain-community
langchain-core
langchain-openai
openai
pydantic
pydantic-core
tiktoken
pytest
pytest-asyncio
pytest-cov
httpx